move simulator object to its own module
